Updated programs
Renamed programs:
game.py -> tamagotchi.py
game_loop.py -> game.py

Finished the ui_interactables.py and implemented to tamagotchi.py. Preparing to merge feature-WIP to working.

v1.8.12
